Moved modules into src. Did a lot of refactoring.
